requirements.txt
paramiko
gpustat
flask
pyyaml